requirements.txt
packaging==23.2
addict==2.4.0
editdistance==0.6.2
einops==0.7.0
fairscale==0.4.0
jsonlines==4.0.0
markdown2==2.4.10
matplotlib==3.7.4
more_itertools==10.1.0
nltk==3.8.1
numpy==1.24.4
opencv_python_headless==4.5.5.64
openpyxl==3.1.2
Pillow==10.1.0
sacrebleu==2.3.2
seaborn==0.13.0
shortuuid==1.0.11
timm==0.9.10
torch==2.3.1
torchvision==0.18.1
protobuf==4.25.0
transformers==4.43.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
sentencepiece==0.1.99
accelerate==0.30.1
socksio==1.0.0
gradio
gradio_client
datasets==2.21.0
deepspeed==0.14.2
peft==0.11.1
sentence-transformers
vllm==0.5.3.post1
wandb==0.17.7
tqdm
av
pyarrow
imageio
imageio-ffmpeg
ego4d
submitit
jupyterlab
tabulate